Sad paths for meals

+++ routes/api/v1/meals.js
@@ -8,6 +8,14 @@ const mealFood = require('../../../models').MealFood;
 router.post('/', async (req, res, next) => {
   res.setHeader('Content-Type', 'application/json')
 
+  if (!req.body.name) {
+    payload = {
+      message: 'Name is required.'
+    }
+    res.status(400).send(payload)
+    return;
+  }
+  
   await meal.create({
     name: req.body.name,
   })
@@ -55,6 +63,13 @@ router.get('/:meal_id/foods', async (req, res, next) => {
     }]
   })
     .then(async meal => {
+      if (!meal) {
+        payload = {
+          message: 'Meal not found.'
+        }
+        res.status(404).send(payload)
+        return;
+      }
       res.status(200).send(meal)
     })
     .catch(async error => {
@@ -74,18 +89,53 @@ router.post('/:meal_id/foods/:id', async (req, res, next) => {
       const getFood = await food.findOne({where: mealFood.foodId})
       const getMeal = await meal.findOne({where: mealFood.mealId})
 
+      if (!getFood) {
+        payload = {
+          message: 'Food not found.'
+        }
+        res.status(404).send(payload)
+        return;
+      }
+
+      if (!getMeal) {
+        payload = {
+          message: 'Meal not found.'
+        }
+        res.status(404).send(payload)
+        return;
+      }
+
       res.status(201).send({
         "message": `Successfully added ${getFood.name} to ${getMeal.name}`
-      })
+      });
     })
     .catch(async error => {
       res.status(500).send({ error })
-    })
-})
+    });
+});
 
 /* DELETE food from meal */
 router.delete('/:meal_id/foods/:id', async (req, res, next) => {
   res.setHeader('Content-type', 'application/json')
+
+  const getFood = await food.findOne({where: {id: req.params.id}})
+  const getMeal = await meal.findOne({where: {id: req.params.meal_id}})
+
+  if (!getFood) {
+    payload = {
+      message: 'Food not found.'
+    }
+    res.status(404).send(payload)
+    return;
+  }
+
+  if (!getMeal) {
+    payload = {
+      message: 'Meal not found.'
+    }
+    res.status(404).send(payload)
+    return;
+  }
 
   await mealFood.destroy({
     where: {
@@ -98,7 +148,7 @@ router.delete('/:meal_id/foods/:id', async (req, res, next) => {
     })
     .catch(async error => {
       res.status(500).send({ error })
-    })
-})
+    });
+});
 
 module.exports = router;
